Restructure plugins into one directory per plugin
