+++ lib/file_scope.rb
