+++ plugins/changed_managed_files/changed_managed_files_inspector.rb
