+++ plugins/changed_managed_files/changed_managed_files_model.rb
@@ -15,16 +15,14 @@
 # To contact SUSE about this file by physical or electronic mail,
 # you may find current contact information at www.suse.com
 
-require_relative "file_scope"
-
-class ConfigFile < Machinery::Object
+class ChangedManagedFile < Machinery::Object
 end
 
-class ConfigFileList < Machinery::Array
-  has_elements class: ConfigFile
+class ChangedManagedFileList < Machinery::Array
+  has_elements class: ChangedManagedFile
 end
 
-class ConfigFilesScope < FileScope
+class ChangedManagedFilesScope < FileScope
   include Machinery::Scope
-  has_property :files, class: ConfigFileList
+  has_property :files, class: ChangedManagedFileList
 end
