+++ plugins/config_files/config_files_inspector.rb
