+++ plugins/config_files/config_files_model.rb
@@ -15,16 +15,14 @@
 # To contact SUSE about this file by physical or electronic mail,
 # you may find current contact information at www.suse.com
 
-require_relative "file_scope"
-
-class ChangedManagedFile < Machinery::Object
+class ConfigFile < Machinery::Object
 end
 
-class ChangedManagedFileList < Machinery::Array
-  has_elements class: ChangedManagedFile
+class ConfigFileList < Machinery::Array
+  has_elements class: ConfigFile
 end
 
-class ChangedManagedFilesScope < FileScope
+class ConfigFilesScope < FileScope
   include Machinery::Scope
-  has_property :files, class: ChangedManagedFileList
+  has_property :files, class: ConfigFileList
 end
