+++ plugins/config_files/config_files_renderer.rb
