+++ plugins/groups/groups_inspector.rb
