+++ plugins/groups/groups_renderer.rb
