+++ plugins/os/os_inspector.rb
