+++ plugins/os/os_renderer.rb
