+++ plugins/packages/packages_inspector.rb
