+++ plugins/packages/packages_renderer.rb
