+++ plugins/patterns/patterns_inspector.rb
