+++ plugins/patterns/patterns_renderer.rb
