+++ plugins/repositories/repositories_inspector.rb
