+++ plugins/repositories/repositories_renderer.rb
