+++ plugins/services/services_inspector.rb
