+++ plugins/unmanaged_files/unmanaged_files_inspector.rb
