+++ plugins/unmanaged_files/unmanaged_files_model.rb
@@ -15,8 +15,6 @@
 # To contact SUSE about this file by physical or electronic mail,
 # you may find current contact information at www.suse.com
 
-require_relative "file_scope"
-
 class UnmanagedFile < Machinery::Object
 end
