+++ plugins/users/users_inspector.rb
